scripts/test-image-requirements.txt
# Requirements for downloading test images
datasets>=2.14.0
pillow>=10.0.0
